services/asr/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
# Install whisper from GitHub to avoid legacy setup.py pkg_resources issue
openai-whisper @ git+https://github.com/openai/whisper.git
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
pydantic==2.7.1
torch==2.2.2
torchaudio==2.2.2
numpy==1.26.4
